pyproject.toml
[project]
name = "berlin-apartment-finder"
version = "0.1.0"
description = "A tool to find and track Berlin apartment listings"
authors = [
    {name = "Matteo Gevi"}
]
readme = "README.md"
requires-python = "^3.9"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "schedule (>=1.2.2,<2.0.0)",
    "selenium (>=4.30.0,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
